Warring cleanup. Exception fixing
Former-commit-id: ac6290a68cf2e6a53d35e8b6a1d2439d0b9b1ca0

+++ PrimeEngine/PrimeEngine/Networking/NetworkSever.cpp
@@ -121,7 +121,7 @@ namespace PrimeEngine { namespace Networking {
 				DisconnectClient(clientId);
 				return;
 			}
-			int w_len = send(_clientSocket[clientId], message, strlen(message), 0);
+			int w_len = send(_clientSocket[clientId], message, (int)strlen(message), 0);
 			if (w_len <= 0)
 			{
 				DisconnectClient(clientId);
